% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmap.R
\name{listmap}
\alias{listmap}
\title{Title}
\usage{
listmap(x, layer = "Topographic", band = 1:3, resample = "bilinear")
}
\arguments{
\item{x}{a raster to define the extent/dimension/projection (later something elses too)}

\item{layer}{name or number (3:14 works)}

\item{resample}{}
}
\value{

}
\description{
Title
}
\examples{
listmap()
}
